pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orgclone"
version = "0.1.0"
description = "Clone entire GitHub orgs or GitLab groups with one command"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
orgclone = "orgclone.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["orgclone"]
